+++ xmtp_mls/src/groups/tests/test_commit_log_remote.rs
@@ -310,7 +310,7 @@ async fn test_publish_commit_log_to_remote() {
         .get_last_cursor_for_originator(
             &alix_group.group_id,
             xmtp_db::refresh_state::EntityKind::CommitLogUpload,
-            Originators::REMOTE_COMMIT_LOG as u32,
+            Originators::REMOTE_COMMIT_LOG,
         )
         .unwrap();
     assert_eq!(published_commit_log_cursor, Cursor::commit_log(0));
@@ -328,7 +328,7 @@ async fn test_publish_commit_log_to_remote() {
         .get_last_cursor_for_originator(
             &alix_group.group_id,
             xmtp_db::refresh_state::EntityKind::CommitLogUpload,
-            Originators::REMOTE_COMMIT_LOG as u32,
+            Originators::REMOTE_COMMIT_LOG,
         )
         .unwrap();
     tracing::info!("{}", published_commit_log_cursor);
@@ -414,7 +414,7 @@ async fn test_download_commit_log_from_remote() {
         .get_last_cursor_for_originator(
             &alix_group.group_id,
             xmtp_db::refresh_state::EntityKind::CommitLogUpload,
-            Originators::REMOTE_COMMIT_LOG as u32,
+            Originators::REMOTE_COMMIT_LOG,
         )
         .unwrap();
     assert_eq!(alix_group_1_cursor, Cursor::commit_log(0));
@@ -472,7 +472,7 @@ async fn test_download_commit_log_from_remote() {
         .get_last_cursor_for_originator(
             &alix_group.group_id,
             xmtp_db::refresh_state::EntityKind::CommitLogUpload,
-            Originators::REMOTE_COMMIT_LOG as u32,
+            Originators::REMOTE_COMMIT_LOG,
         )
         .unwrap();
 
@@ -482,7 +482,7 @@ async fn test_download_commit_log_from_remote() {
         alix_group_1_cursor,
         Cursor::new(
             alix_group1_publish_result_upload_cursor as u64,
-            Originators::REMOTE_COMMIT_LOG as u32
+            Originators::REMOTE_COMMIT_LOG
         )
     );
